Refactor: Change `requiredQuantity` from Double to Int
This commit updates the data type for an ingredient's required quantity from `Double` to `Int` across the domain model, data entity, view model, and UI components.

**Changes:**

*   **`RecipeIngredient.kt` & `RecipeIngredientEntity.kt`:**
    *   Changed the `requiredQuantity` property from `Double` to `Int`.

*   **`AddRecipeScreenViewModel.kt`:**
    *   Updated `onRecipeIngredientChangeQuantity` to accept an `Int` instead of a `Double`.
    *   Adjusted the logic for adding a new recipe ingredient to use the integer quantity directly without converting it to a `Double`.

*   **`AddRecipeScreen.kt`:**
    *   Modified function signatures and calls to use `Int` for quantity values.
    *   Removed the `.toDouble()` and `.toInt()` conversions in the `TextField`'s `onValueChange` and the `IngredientCard`'s `quantity` parameter, respectively.

+++ app/src/main/java/diomaxius/whattocookwith/data/model/recipe/RecipeIngredientEntity.kt
@@ -26,7 +26,7 @@ import diomaxius.whattocookwith.data.model.ingredient.IngredientEntity
 data class RecipeIngredientEntity(
     val recipeId: Long,
     val ingredientName: String,
-    val requiredQuantity: Double,
+    val requiredQuantity: Int,
     val unit: String,
     val optional: Boolean = false,
 )
