+++ app/src/main/java/diomaxius/whattocookwith/domain/model/RecipeIngredient.kt
@@ -3,7 +3,7 @@ package diomaxius.whattocookwith.domain.model
 data class RecipeIngredient(
     val recipeId: Long,
     val ingredientName: String,
-    val requiredQuantity: Double,
+    val requiredQuantity: Int,
     val unit: String,
     val optional: Boolean
 )
